+++ dashboard/src/features/projects/rate-limiting/settings/components/RunServiceLimitingForm/index.ts
@@ -1,1 +1,1 @@
-export { default as RunFunctionLimitingForm } from './RunFunctionLimitingForm';
+export { default as RunServiceLimitingForm } from './RunServiceLimitingForm';
